Add blog RSS feed
The riff homepage includes a link to an Atom feed at /feed.xml, however,
this feed is not generated by Jekyll. By adding the jekyll-feed plugin,
a feed will automatically be generated from posts.

The content of the first blog post has moved from `_pages` to `_posts`
in order to be detected by the plugin and added to the feed. The content
and public URLs for the post have not changed.