Spike: Invokers listing
Adds an /invokers section of the website.

Each invoker is listed on the page and also has a dedicated page. The
listing includes a link to the details page and the comamnd to install
the most recent release (or the development version). The detail page
includes instructions on how to install the most recent release, the
current development build, historical releases and a summary of the
invoker.

Each invoker must be saved in the `_invokers` directory and include a
header like:

```
---
name: command
title: "Command Function Invoker"
author: projectriff
href: https://github.com/projectriff/command-function-invoker
development:
  href: https://github.com/projectriff/command-function-invoker/raw/master/command-invoker.yaml
releases:
- version: v0.0.6
  href: https://github.com/projectriff/command-function-invoker/raw/v0.0.6/command-invoker.yaml
  date: 2000-01-01
- version: v0.0.5
  href: https://github.com/projectriff/command-function-invoker/raw/v0.0.5/command-invoker.yaml
  date: 1999-01-01
- version: v0.0.4
  href: https://github.com/projectriff/command-function-invoker/raw/v0.0.4/command-invoker.yaml
  date: 1998-01-01
---
```

Content of the invoker file will show on the detail page.

The styling is abmismal and needs love.

Issue: #14